requirements.txt
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
nltk==3.3
numpy==1.15.4
pycparser==2.19
PySoundFile==0.9.0.post1
python-dotenv==0.9.1
requests==2.20.0
six==1.11.0
sounddevice==0.3.12
urllib3==1.24.1
Werkzeug==0.14.1
